+++ project/routine/controllers/item_controller.py
@@ -1,5 +1,5 @@
-from flask import render_template, flash, redirect, url_for
-from flask_login import login_required
+from flask import render_template, flash, redirect, url_for, abort
+from flask_login import login_required, current_user
 
 from .. import item_bp
 from ..forms import NewItemForm
@@ -15,6 +15,8 @@ def add(routine_id: int):
     form = NewItemForm()
     routine = routine_service.get_by_id(routine_id)
     try:
+        if routine.created_by != current_user.id:
+            abort(403)
         if form.validate_on_submit():
             item = item_service.add(
                 routine_type=routine.type.value,
@@ -37,7 +39,11 @@ def remove(item_id: int):
     item = item_service.find_by_id(item_id)
     routine_id = item.routine_id
     try:
-        item_service.delete(item_id)
+        routine = routine_service.get_by_id(routine_id)
+        if routine:
+            if routine.created_by != current_user.id:
+                abort(403)
+            item_service.delete(item_id)
     except Exception as e:
         flash(str(e))
     return redirect(url_for('routine.update', routine_id=routine_id))
@@ -50,6 +56,8 @@ def detail(item_id:int):
         item = item_service.find_by_id(item_id)
         if not item:
             raise Exception('Item não encontrado.')
+        if item.routine.created_for != current_user.id and item.routine.created_by != current_user.id:
+            abort(403)
         return render_template('item_routine/detail_item.html', item=item)
     except Exception as e:
         flash(str(e))
