+++ project/user/services/user_service.py
@@ -1,4 +1,5 @@
 from flask import abort
+from flask_login import current_user
 
 from .. import User
 from ..repositories import UserRepository
@@ -22,4 +23,9 @@ class UserService:
         return self.user_repository.update_email(user, email)
 
     def delete(self, user: User):
-        self.user_repository.delete(user)
+        if user.profile.role == 'admin' and current_user.role != 'god':
+            raise Exception('Você não pode excluir admins, quem você pensa que é?')
+        elif user.profile.id != current_user.id and current_user.role == 'user':
+                raise Exception('Você não pode excluir contas alheias')
+        else:
+            self.user_repository.delete(user)
